finish the achievement and goal logic

+++ Achievement-board/src/main/java/com/example/achievementboard/repos/AchievementRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AchievementRepository extends JpaRepository<Achievement , Integer> {
+public interface AchievementRepository extends JpaRepository<Achievement , Long> {
     List<Achievement> findAllByUser(User user);
 }
