+++ Achievement-board/src/main/java/com/example/achievementboard/repos/GoalRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface GoalRepository extends JpaRepository<Goal, Integer> {
+public interface GoalRepository extends JpaRepository<Goal, Long> {
     List<Goal> findAllByUser(User user);
     List<Goal> findAllByUserOrderByEndDate(User user);
 }
