+++ Achievement-board/src/main/java/com/example/achievementboard/service/BaseService.java
@@ -7,7 +7,7 @@ public interface BaseService<T> {
 
     void saveAll(List<T> build);
 
-    T getById(Integer id);
+    T getById(Long id);
 
     void save(T entity);
 }
